requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2

# Database
pymongo==4.6.0

# Optional database connectors (install only if needed)
# mysql-connector-python==8.0.33
# psycopg2-binary==2.9.6
